+++ src/cashew/parser.cpp
@@ -95,7 +95,7 @@ int MAX_OPERATOR_SIZE = 3;
 
 std::vector<OperatorClass> operatorClasses;
 
-static std::vector<std::unordered_map<IString, int>> precedences; // op, type => prec
+static std::vector<std::unordered_map<IString, size_t>> precedences; // op, type => prec
 
 struct Init {
   Init() {
@@ -126,7 +126,7 @@ struct Init {
 
 Init init;
 
-int OperatorClass::getPrecedence(Type type, IString op) {
+size_t OperatorClass::getPrecedence(Type type, IString op) {
   return precedences[type][op];
 }
