+++ src/cashew/parser.h
@@ -116,7 +116,7 @@ struct OperatorClass {
 
   OperatorClass(const char* o, bool r, Type t) : ops(o), rtl(r), type(t) {}
 
-  static int getPrecedence(Type type, IString op);
+  static size_t getPrecedence(Type type, IString op);
   static bool getRtl(int prec);
 };
